feat: move GroupProfileImage component to common
