+++ src/components/common/GroupProfileImage/GroupProfileImage.stories.tsx
