+++ src/components/common/GroupProfileImage/index.ts
