+++ src/components/common/Profile/index.ts
@@ -1,6 +1,3 @@
-import GroupProfileImage from './GroupProfileImage';
 import Profile from './Profile';
-
-export { GroupProfileImage };
 
 export default Profile;
